@@ -6,7 +6,8 @@ package com.adityagupta.nxtvisioncallrecorder.utils.typeface;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.v4.util.SimpleArrayMap;
+
+import androidx.collection.SimpleArrayMap;
 
 /*
     Each call to Typeface.createFromAsset will load a new instance of the typeface into memory,
