@@ -6,7 +6,8 @@ import android.view.accessibility.AccessibilityEvent;
 public class MyAccessibilityService extends AccessibilityService {
 
     @Override
-    public void onAccessibilityEvent(AccessibilityEvent event) { }
+    public void onAccessibilityEvent(AccessibilityEvent event) {
+    }
 
     @Override
     public void onInterrupt() {
