@@ -6,11 +6,13 @@ import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 import android.telephony.TelephonyManager;
 import android.util.Log;
+
 import com.adityagupta.nxtvisioncallrecorder.application.ApplicationActivity;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.JsonObjectRequest;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -22,22 +24,22 @@ public class Preferences {
         return device_id;
     }
 
-    public static void checkLicense(Context context){
-        Log.e("Lincense Check","Checking");
+    public static void checkLicense(Context context) {
+        Log.e("Lincense Check", "Checking");
         final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
 
         JsonObjectRequest req = new JsonObjectRequest("https://raw.githubusercontent.com/adityastic/LicensingRepo/master/NxtRecorder.json", null,
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject response) {
-                        Log.e("Record Enabled",response.toString());
+                        Log.e("Record Enabled", response.toString());
                         try {
-                            prefs.edit().putBoolean("allowed",response.getBoolean("Record")).apply();
+                            prefs.edit().putBoolean("allowed", response.getBoolean("Record")).apply();
                         } catch (JSONException e) {
-                            prefs.edit().putBoolean("allowed",false).apply();
+                            prefs.edit().putBoolean("allowed", false).apply();
                             e.printStackTrace();
                         }
-                        Log.e("Saved Value",prefs.getBoolean("allowed",false) +"");
+                        Log.e("Saved Value", prefs.getBoolean("allowed", false) + "");
                     }
                 }, new Response.ErrorListener() {
             @Override
