@@ -11,8 +11,9 @@ public class TypefaceUtil {
     /**
      * Using reflection to override default typeface
      * NOTICE: DO NOT FORGET TO SET TYPEFACE FOR APP THEME AS DEFAULT TYPEFACE WHICH WILL BE OVERRIDDEN
-     * @param context to work with assets
-     * @param defaultFontNameToOverride for example "monospace"
+     *
+     * @param context                    to work with assets
+     * @param defaultFontNameToOverride  for example "monospace"
      * @param customFontFileNameInAssets file name of the font from assets
      */
     public static void overrideFont(Context context, String defaultFontNameToOverride, String customFontFileNameInAssets) {
@@ -23,7 +24,7 @@ public class TypefaceUtil {
             defaultFontTypefaceField.setAccessible(true);
             defaultFontTypefaceField.set(null, customFontTypeface);
         } catch (Exception e) {
-            Log.e("Exception","Can not set custom font " + customFontFileNameInAssets + " instead of " + defaultFontNameToOverride);
+            Log.e("Exception", "Can not set custom font " + customFontFileNameInAssets + " instead of " + defaultFontNameToOverride);
         }
     }
 }
